+++ __app/component/AutoFillOtp/AutoFillOtp.js
@@ -1,4 +1,14 @@
-import { handleError, handleSuccess } from '../services/handlerService';
+import React from 'react';
+import PropTypes from 'prop-types';
+import { handleSuccess, handleError, handleLoading } from '../services/handlerService';
+import Wrapper from '../Wrapper/Wrapper';
+
+const failureMsgDefault = {
+  unSupported: 'AutoFillOtp is not supporting in your device',
+  error: 'Unable to fetch details from AutoFillOtp',
+};
+
+const isBrowserSupport = () => globalThis.OTPCredential;
 
 const abortAutoFill = (abort, time) => {
   setTimeout(() => {
@@ -7,36 +17,67 @@ const abortAutoFill = (abort, time) => {
   }, time * 60 * 1000);
 };
 
-const defaultProps = {
-  successCb: (() => {}),
-  failureCb: (() => {}),
-  successMsg: 'OTP autofilled successfully',
-  failureMsg: {
-    UN_SUPPORTED_FEATURE: 'Your device is not supporting AutofillOTP',
-    ERROR: '',
-  },
-};
-function AutoFillOtp(props = {}) {
-  const successCb = props.successCb || defaultProps.successCb;
-  const failureCb = props.failureCb || defaultProps.failureCb;
-  const successMsg = props.successMsg || defaultProps.successMsg;
-  const failureMsg = { ...defaultProps.failureMsg, ...props.failureMsg };
+const autoFillOtp = ({
+  successCb = () => {},
+  failureCb = () => {},
+  loadingCb = () => {},
+  successMsg = 'OTP autofilled successfully',
+  failureMsg: failureMsgProps = { ...failureMsgDefault },
+} = {}) => {
+  const failureMsg = { ...failureMsgDefault, ...failureMsgProps };
 
-  if (AutoFillOtp.isBrowserSupport()) {
-    const abort = new AbortController();
-    abortAutoFill(abort, 3);
-    navigator.credentials.get({
-      otp: { transport: ['sms'] },
-      signal: abort.signal,
-    }).then((otp) => {
-      const { code } = otp;
-      handleSuccess({ msgType: 'SUCCESSFUL', msg: successMsg, successCb, data: code });
-    }).catch((error) => handleError({ msgType: 'ERROR', msg: error, failureCb }));
-  } else {
-    return handleError({ msgType: 'UN_SUPPORTED_FEATURE', msg: failureMsg.unSupported, failureCb });
-  }
+  const init = () => {
+    if (isBrowserSupport()) {
+      handleLoading({ loadingCb });
+
+      const abort = new AbortController();
+      abortAutoFill(abort, 3);
+      navigator.credentials.get({
+        otp: { transport: ['sms'] },
+        signal: abort.signal,
+      }).then((otp) => {
+        const { code } = otp;
+        handleSuccess({ msgType: 'SUCCESSFUL', msg: successMsg, successCb, data: code });
+      }).catch((error) => handleError({ msgType: 'ERROR', msg: error, failureCb }));
+    } else {
+      return handleError({ msgType: 'UN_SUPPORTED_FEATURE', msg: failureMsg.unSupported, failureCb });
+    }
+    return true;
+  };
+
+  init();
+};
+
+function AutoFillOtp({
+  children,
+  successCb,
+  failureCb,
+  loadingCb,
+  successMsg,
+  failureMsg,
+}) {
+  return React.Children.map(children || 'AutoFillOtp', (child) => React.cloneElement(typeof child === 'string' ? <span>{child}</span> : child, {
+    onClick: () => autoFillOtp({
+      successCb,
+      failureCb,
+      loadingCb,
+      successMsg,
+      failureMsg,
+    }),
+  }));
 }
 
-AutoFillOtp.isBrowserSupport = () => globalThis.OTPCredential;
+AutoFillOtp.propTypes = {
+  showForever: PropTypes.bool,
+  successCb: PropTypes.func,
+  failureCb: PropTypes.func,
+  loadingCb: PropTypes.func,
+  successMsg: PropTypes.string,
+  failureMsg: PropTypes.object,
+};
 
-export default AutoFillOtp;
+const WAutoFillOtp = Wrapper(AutoFillOtp, isBrowserSupport);
+
+export { autoFillOtp, WAutoFillOtp as AutoFillOtp };
+
+export default WAutoFillOtp;