+++ __app/component/CopyToClipboard/CopyToClipboard.js
@@ -4,42 +4,63 @@ import { handleSuccess, handleError, handleLoading } from '../services/handlerSe
 import Wrapper from '../Wrapper/Wrapper';
 
 const failureMsgDefault = {
-  unSupported: 'Copy To ClipBoard is not supporting in your device',
-  error: 'Unable to copy',
+  unSupported: 'CopyToClipboard is not supporting in your device',
+  error: 'Unable To Copy',
 };
 
-function CopyToClipboard({
-  successCb,
-  failureCb,
-  loadingCb,
-  successMsg,
-  failureMsg: failureMsgProps,
-  children,
+const isBrowserSupport = () => globalThis.navigator?.clipboard;
+
+const copyToClipboard = ({
+  successCb = () => {},
+  failureCb = () => {},
+  loadingCb = () => {},
+  successMsg = 'Copied Successfully',
+  failureMsg: failureMsgProps = { ...failureMsgDefault },
   elementToBeCopy,
-}) {
+} = {}) => {
   const failureMsg = { ...failureMsgDefault, ...failureMsgProps };
 
-  const copyText = () => {
-    if (CopyToClipboard.isBrowserSupport()) {
+  const init = () => {
+    if (isBrowserSupport()) {
       handleLoading({ loadingCb });
+
+      // Your Code will start from here
       navigator.clipboard.writeText(elementToBeCopy).then(() => {
         handleSuccess({ msgType: 'SUCCESSFUL', msg: successMsg, successCb, data: elementToBeCopy });
       }).catch((error) => handleError({ msgType: 'ERROR', msg: failureMsg.error || JSON.stringify(error), failureCb }));
+      // Your Code will end here
     } else {
       return handleError({ msgType: 'UN_SUPPORTED_FEATURE', msg: failureMsg.unSupported, failureCb });
     }
     return true;
   };
 
-  return React.Children.map(children || 'Copy', (child) => React.cloneElement(typeof child === 'string' ? <span>{child}</span> : child, {
-    onClick: copyText,
-    onKeyDown: copyText,
+  init();
+};
+
+function CopyToClipboard({
+  children,
+  successCb,
+  failureCb,
+  loadingCb,
+  successMsg,
+  failureMsg,
+  ...props
+}) {
+  return React.Children.map(children || 'CopyToClipboard', (child) => React.cloneElement(typeof child === 'string' ? <span>{child}</span> : child, {
+    onClick: () => copyToClipboard({
+      successCb,
+      failureCb,
+      loadingCb,
+      successMsg,
+      failureMsg,
+      ...props,
+    }),
   }));
 }
 
-CopyToClipboard.isBrowserSupport = () => navigator.clipboard && true;
-
 CopyToClipboard.propTypes = {
+  showForever: PropTypes.bool,
   successCb: PropTypes.func,
   failureCb: PropTypes.func,
   loadingCb: PropTypes.func,
@@ -47,12 +68,8 @@ CopyToClipboard.propTypes = {
   failureMsg: PropTypes.object,
 };
 
-CopyToClipboard.defaultProps = {
-  successCb: () => {},
-  failureCb: () => {},
-  loadingCb: () => {},
-  successMsg: 'Copied Successfully',
-  failureMsg: { ...failureMsgDefault },
-};
+const WCopyToClipboard = Wrapper(CopyToClipboard, isBrowserSupport);
 
-export default Wrapper(CopyToClipboard);
+export { copyToClipboard, WCopyToClipboard as CopyToClipboard };
+
+export default WCopyToClipboard;