+++ __app/component/Scanner/Scanner.js
@@ -17,6 +17,8 @@ const failureMsgDefault = {
   unableToScan: 'Unable to scan',
 };
 
+const isBrowserSupport = () => globalThis.navigator?.mediaDevices && globalThis.BarcodeDetector;
+
 function Scanner({
   successCb,
   failureCb,
@@ -94,6 +96,7 @@ function Scanner({
     video.style.top = '0';
     video.style.left = '0';
     video.style.objectFit = 'fill';
+    video.style.transform = 'rotateY(180deg)';
     list = document.getElementById(id);
     list.before(video);
   };
@@ -151,7 +154,7 @@ function Scanner({
   };
 
   const handleBrowserSupport = () => {
-    if (Scanner.isBrowserSupport()) {
+    if (isBrowserSupport()) {
       facingMode = cameraType === 'back' ? 'environment' : 'user';
       handleLoading({ loadingCb });
 
@@ -172,7 +175,7 @@ function Scanner({
     };
   }, []);
 
-  return isBrowser && Scanner.isBrowserSupport() && (
+  return isBrowser && isBrowserSupport() && (
     <div id="scanner">
       <div id="camera" />
       {
@@ -190,8 +193,6 @@ function Scanner({
     </div>
   );
 }
-
-Scanner.isBrowserSupport = () => navigator?.mediaDevices && globalThis.BarcodeDetector;
 
 Scanner.propTypes = {
   successCb: PropTypes.func,
@@ -213,4 +214,8 @@ Scanner.defaultProps = {
   cameraType: 'back',
 };
 
-export default Wrapper(Scanner);
+const WScanner = Wrapper(Scanner, isBrowserSupport);
+
+export { WScanner as Scanner };
+
+export default Scanner;