+++ __app/component/WhatsappShare/WhatsappShare.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { handleSuccess, handleError } from '../services/handlerService';
+import { handleSuccess, handleError, handleLoading } from '../services/handlerService';
 import Wrapper from '../Wrapper/Wrapper';
 
 const failureMsgDefault = {
@@ -8,66 +8,82 @@ const failureMsgDefault = {
   badRequest: 'msg is missing',
   error: 'Unable to fetch details from WhatsappShare',
 };
+const isBrowserSupport = () => globalThis;
 
-function WhatsappShare({
-  successCb,
-  failureCb,
-  successMsg,
-  failureMsg: failureMsgProps,
-  children,
+const whatsAppShare = ({
+  successCb = () => {},
+  failureCb = () => {},
+  loadingCb = () => {},
+  successMsg = 'WhatsappShare details fetch Successfully!!',
+  failureMsg: failureMsgProps = { ...failureMsgDefault },
   mobile,
   msg,
-}) {
+} = {}) => {
   const failureMsg = { ...failureMsgDefault, ...failureMsgProps };
 
-  const getWhatsappShare = () => {
-    if (navigator.userAgent.match(/iPhone|Android/i)) {
-      if (mobile) {
-        window.location.href = `https://wa.me/${mobile}/?text=${msg}`;
-        handleSuccess({ msgType: 'SUCCESSFUL', msg: successMsg, successCb, data: msg });
-      } else if (msg) {
-        window.location.href = `https://wa.me/?text=${msg}`;
-        handleSuccess({ msgType: 'SUCCESSFUL', msg: successMsg, successCb, data: msg });
-      } else {
-        return handleError({ msgType: 'BAD_REQUEST', msg: failureMsg.badRequest, failureCb });
-      }
-    } else {
-      window.location.href = `https://web.whatsapp.com/send?text=${msg}&phone=${mobile}`;
-    }
+  const init = () => {
+    if (isBrowserSupport()) {
+      handleLoading({ loadingCb });
 
+      // Your Code will start from here
+      if (navigator.userAgent.match(/iPhone|Android/i)) {
+        if (mobile) {
+          window.location.href = `https://wa.me/${mobile}/?text=${msg}`;
+          handleSuccess({ msgType: 'SUCCESSFUL', msg: successMsg, successCb, data: msg });
+        } else if (msg) {
+          window.location.href = `https://wa.me/?text=${msg}`;
+          handleSuccess({ msgType: 'SUCCESSFUL', msg: successMsg, successCb, data: msg });
+        } else {
+          return handleError({ msgType: 'BAD_REQUEST', msg: failureMsg.badRequest, failureCb });
+        }
+      } else {
+        window.location.href = `https://web.whatsapp.com/send?text=${msg}&phone=${mobile}`;
+      }
+
+      // Your Code will end here
+    } else {
+      return handleError({ msgType: 'UN_SUPPORTED_FEATURE', msg: failureMsg.unSupported, failureCb });
+    }
     return true;
   };
 
-  return (
-    <div>
-      {React.Children.map(children || 'Whatsapp Share', (child) => React.cloneElement(
-        typeof child === 'string' ? <span>{child}</span> : child,
-        {
-          onClick: getWhatsappShare,
-        },
-      ))}
-    </div>
-  );
+  init();
+};
+
+function WhatsAppShare({
+  children,
+  successCb,
+  failureCb,
+  loadingCb,
+  successMsg,
+  failureMsg,
+  ...props
+}) {
+  return React.Children.map(children || 'WhatsAppShare', (child) => React.cloneElement(typeof child === 'string' ? <span>{child}</span> : child, {
+    onClick: () => whatsAppShare({
+      successCb,
+      failureCb,
+      loadingCb,
+      successMsg,
+      failureMsg,
+      ...props,
+    }),
+  }));
 }
 
-WhatsappShare.isBrowserSupport = () => globalThis && true;
-
-WhatsappShare.propTypes = {
+WhatsAppShare.propTypes = {
+  showForever: PropTypes.bool,
   successCb: PropTypes.func,
   failureCb: PropTypes.func,
+  loadingCb: PropTypes.func,
   successMsg: PropTypes.string,
   failureMsg: PropTypes.object,
   mobile: PropTypes.number,
   msg: PropTypes.string,
 };
 
-WhatsappShare.defaultProps = {
-  successCb: () => {},
-  failureCb: () => {},
-  successMsg: 'WhatsappShare details fetch Successfully',
-  failureMsg: { ...failureMsgDefault },
-  mobile: '',
-  msg: '',
-};
+const WWhatsAppShare = Wrapper(WhatsAppShare, isBrowserSupport);
 
-export default Wrapper(WhatsappShare);
+export { whatsAppShare, WWhatsAppShare as WhatsAppShare };
+
+export default WWhatsAppShare;